Added general bedtime angle data, fixed issue with categorical interpolation
