+++ data/basic/summary_engineered_inclbooleans.csv
@@ -19,4 +19,5 @@ est_age_days,float64,151,184,140,7313.0,10464.0,8714.80795,0.0,"['8589.0', '8178
 sleep_duration_to_10am,float64,222,42,55,-0.61667,23.0,10.03664,0.0,"['10.0', '9.833333333333334', '9.5']"
 bedtime_sin,float64,204,78,44,-0.96593,0.96593,0.15456,0.0,"['0.0', '0.043619387365336', '0.13052619222005157']"
 bedtime_cos,float64,204,78,41,0.25882,1.0,0.89662,0.0,"['1.0', '0.9990482215818578', '0.9914448613738104']"
-bedtime_anglenorm,float64,222,42,55,-3.14159,3.03251,0.07532,0.0,"['0.0', '0.04363323129985824', '0.1308996938995747']"
+bedtime_anglenorm,float64,204,78,44,-1.309,1.309,0.1781,0.0,"['0.0', '0.04363323129985824', '0.1308996938995747']"
+bedtime_angle,float64,204,78,44,0.0,6.27882,2.3957,0.0,"['0.0', '0.04363323129985824', '0.1308996938995747']"
