+++ data/basic/summary_knn_knn.csv
@@ -1,22 +1,23 @@
 Column,Data Type,Non-Null Count,Missing Count,Unique Values,Min,Max,Mean,Outlier Count (IQR),Example Values
 timestamp,category,243,0,5,,,,,"['2025-04-01 12:19:00', '2025-04-01 12:18:00', '2025-04-01 12:17:00']"
 program,category,243,0,8,,,,,"['AI', 'Bioinformatics', 'Business']"
-ml_course,category,243,0,3,,,,,"['0.0', '1.0', 'nan']"
-ir_course,category,243,0,3,,,,,"['0.0', '1.0', 'nan']"
-stats_course,category,243,0,3,,,,,"['1.0', '0.0', 'nan']"
-db_course,category,243,0,3,,,,,"['1.0', 'nan', '0.0']"
+ml_course,category,243,0,2,,,,,"['0.0', '1.0']"
+ir_course,category,243,0,2,,,,,"['0.0', '1.0']"
+stats_course,category,243,0,2,,,,,"['1.0', '0.0']"
+db_course,category,243,0,2,,,,,"['1.0', '0.0']"
 gender,category,243,0,3,,,,,"['female', 'male', 'other']"
-used_chatgpt,category,243,0,3,,,,,"['1.0', 'nan', '0.0']"
+used_chatgpt,category,243,0,2,,,,,"['1.0', '0.0']"
 birthday,category,243,0,63,,,,,"['2001-04-10', 'NaT', '2002-05-05']"
-room_estimate,float64,243,0,67,5.0,600.0,342.34705,0.0,"['200.0', '300.0', '280.0']"
-stress_level,float64,243,0,67,-10.0,110.0,46.65815,0.0,"['20.0', '50.0', '85.0']"
+room_estimate,float64,243,0,67,5.0,600.0,341.97805,0.0,"['200.0', '300.0', '280.0']"
+stress_level,float64,243,0,68,-10.0,110.0,46.45376,0.0,"['20.0', '50.0', '85.0']"
 sports_hours,float64,243,0,23,0.0,15.0,5.40439,0.0,"['8.0', '4.0', '7.0']"
-random_number,float64,243,0,120,-1.0,888.0,64.20061,0.0,"['9.0', '87.0', '93.0']"
+random_number,float64,243,0,117,-1.0,888.0,64.56275,0.0,"['9.0', '87.0', '93.0']"
 bedtime,category,243,0,56,,,,,"['00:00:00', '00:10:00', '00:30:00']"
 good_day_1,category,243,0,174,,,,,"['Good sleep', 'Reading', 'That I feel content and happy']"
 good_day_2,category,243,0,199,,,,,"['Friends', 'F1', 'Doing something ']"
-est_age_days,float64,243,0,215,7313.0,10464.0,8786.67215,0.0,"['8589.0', '8178.0', '8376.0']"
-sleep_duration_to_10am,float64,243,0,66,-0.61667,23.0,10.0409,0.0,"['10.0', '9.833333333333334', '9.5']"
-bedtime_sin,float64,243,0,77,-0.96593,0.96593,0.18275,0.0,"['0.0', '0.043619387365336', '0.13052619222005157']"
-bedtime_cos,float64,243,0,57,0.25882,1.0,0.89307,0.0,"['1.0', '0.9990482215818578', '0.9914448613738104']"
-bedtime_anglenorm,float64,243,0,77,-1.309,1.309,0.20774,0.0,"['0.0', '0.04363323129985824', '0.1308996938995747']"
+est_age_days,float64,243,0,214,7313.0,10464.0,8798.00686,0.0,"['8589.0', '8178.0', '8376.0']"
+sleep_duration_to_10am,float64,243,0,67,-0.61667,23.0,10.02961,0.0,"['10.0', '9.833333333333334', '9.5']"
+bedtime_sin,float64,243,0,72,-0.96593,0.96593,0.18275,0.0,"['0.0', '0.043619387365336', '0.1305261922200515']"
+bedtime_cos,float64,243,0,59,0.25882,1.0,0.89307,0.0,"['1.0', '0.9990482215818578', '0.9914448613738104']"
+bedtime_anglenorm,float64,243,0,75,-1.309,1.309,0.20774,0.0,"['0.0', '0.04363323129985824', '0.13089969389957462']"
+bedtime_angle,float64,243,0,72,0.0,6.27882,2.25042,0.0,"['0.0', '0.04363323129985824', '0.13089969389957462']"
